% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map with accident locations}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Integer value between 0 and 50.}

\item{year}{The year}
}
\value{
a map object. Prints the map as a side-effect.
}
\description{
This function creates a map of the accident locations for a given state and year.
  Returns an error message of Invalid year: <value> if the year value
  is not between 2013-2015. Returns an error message of Invalid STATE number:
  if the state number is not one of the States included in the data.
  Returns an error of no accidents to plot if there are no accidents for the
  specified state number and year.
}
\examples{
\dontrun{
fars_map_state(1, 2014)
fars_map_state(20, 2015)
}

}

