% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads a FARS file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The desired filename to be read.}
}
\value{
object of class tbl_df
}
\description{
This function creates a tbl_df object from the specified filename. Prints the resulting tbl_df a side-effect.
  An error message is returned if the file does not exist.
}
\examples{
\dontrun{
fars_read("accident_2015.csv.bz2")
}

}

