% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple FARS files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years.}
}
\value{
A tbl_df with the month and year for each of the years that were specified in the function argument.
  Prints the tbl_df as a side-effect.
}
\description{
This function creates a tbl_df object from multiple files using a list of years as the function argument.
  Prints the resulting tbl_df a side-effect. An error message Invalid Year: <value> is returned if
  the provided year is not between 2013 and 2015.
}
\examples{
\dontrun{
fars_read_years(2013)
fars_read_years(c(2013, 2014, 2015))
}

}

