% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates a FARS file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year}
}
\value{
string in the form of "accident_YYYY.csv.bz2"
}
\description{
This function returns a properly-formatted file name using the year as a function argument.
  An error will occur if the year parameter cannot be coerced into an integer.
}
\examples{
\dontrun{
make_filename(2016)
}

}

