% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_map_state}
\alias{fars_map_state}
\title{MapByState}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State number as defined in the csv file - should be an integer
between 1 and 50}

\item{year}{Year to be queried for number of accidents - should be an integer
for the years 2013 through 2015}
}
\value{
A map of a state, with points notating accidents for that year
}
\description{
Map Traffic Accidents by State and Year
}
\details{
A function that reads data from the US National Highway Traffic Safety
Administration's Fatality Analysis Reporting System and maps the number of fatalities
for one state for a particular year

Will report an error if there is an invalid state number or if there are
   no accidents in that state in the specified year or if multiple states or years
   are chosen
}
\examples{
   fars_map_state(12, 2015)
}
