% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read}
\alias{fars_read}
\title{ReadData}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Name of file to be input into the function. This is passed
from the make_filename function (and can have the .bz2 extension)
input should be a text string}
}
\value{
This function returns a data frame read from the input file
}
\description{
Read data
}
\details{
Takes the filename string and attempts to open the csv file as a data frame
Not user-facing

will throw an error if the filename doesn't exist in the path
}
\examples{
   fars_read("accident_2014.csv.bz2")
}
