% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read_years}
\alias{fars_read_years}
\title{SelectYearsFromData}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Parameter is passed from the fars_summarize_years function - should be an integer}
}
\value{
Selects all of the months for the year being requested
}
\description{
Gather data
}
\details{
Compiles all of the months data for the year being selected by the user
Not user-facing

Will throw an error if the year being passed in not in the filenames available
}
\examples{
   fars_read_years(2015)
}
