% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{SummarizeByYears}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{User inputs the year needed - should be an integer
between 2013 and 2015}
}
\value{
Returns a table with number of accidents, by month, in the
   selected year
}
\description{
Summarize traffic fatalities by month and year
}
\details{
A function that can summarize multiple years, by month. It takes the user input
of a year or multiple years and produces a table of traffic fatalities by month
}
\examples{
   fars_summarize_years(2015)
}
