% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Parameter is passed from the fars_read_years function
input that is passed should be able to be converted to an integer}
}
\value{
Creates a string for the filename
}
\description{
Creates a filename
}
\details{
Take the year input by the user and creates a string to map to
the file to be read - for that year.
Not user-facing
}
\examples{
   make_filename(2015)
}
